Trait retina_datatypes::PacketList

pub trait PacketList {
    // Required methods
    fn new(first_pkt: &L4Pdu) -> Self;
    fn track_packet(&mut self, pdu: &L4Pdu, reassembled: bool);
    fn clear(&mut self);
}

Trait for a datatype that is built from a list of raw packets.

Required Methods

fn new(first_pkt: &L4Pdu) -> Self

Initialize internal data; called once per connection. Note first_pkt will also be delivered to update.

fn track_packet(&mut self, pdu: &L4Pdu, reassembled: bool)

New packet in connection received (or reassembled, if reassembled=true) Note this may be invoked both pre- and post-reassembly; types should check reassembled to avoid double-counting.

fn clear(&mut self)

Clear internal data; called if connection no longer matches filter that requires the Tracked type.

Object Safety

This trait is not object safe.