retina_datatypes
Module typedefs

New datatypes are defined in this module. Newly-defined datatypes must be added to the DATATYPES map in this module.

Structs

DATATYPES

To add a datatype, add it to the DATATYPES map This is read by the filtergen crate to generate code

DIRECTLY_TRACKED

To avoid copying, the Tracked structure in the framework – built at compile time – will track certain generic, raw datatypes if a subset of subscriptions require them.

Type Aliases

FilterStr

The string literal representing a matched filter.

SessionList

A list of all sessions (zero-copy) parsed in the connection.