retina_datatypes::typedefs
Struct DIRECTLY_TRACKED

pub struct DIRECTLY_TRACKED { /* private fields */ }

To avoid copying, the Tracked structure in the framework – built at compile time – will track certain generic, raw datatypes if a subset of subscriptions require them.

For example: buffering packets may be required as a pre-match action for a packet-level datatype; it may also be required if one or more subscriptions request a connection-level packet list. Rather than maintaining these lists separately – one for filtering and one for delivery – the tracked packets are stored once.

Core ID is a special case, as it cannot be derived from connection, session, or packet data. It is simpler to define it as a directly tracked datatype.

The directly tracked datatypes are SessionList and CoreId